♻️ 컴포넌트 및 로직 분리
