+++ src/components/common/modal/detail-cards/AuthorSection.module.css
@@ -1,0 +1,55 @@
+.author-section {
+  position: absolute;
+  top: 80px;
+  right: 32px;
+
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+
+  width: 200px;
+
+  padding: 14px 16px;
+
+  border: 1px solid var(--gray-medium);
+  border-radius: 8px;
+}
+
+.author-title {
+  margin-bottom: 6px;
+
+  font-size: 12px;
+  font-weight: 600;
+  line-height: 20px;
+}
+
+.author-content {
+  font-size: 14px;
+  line-height: 24px;
+  color: var(--black-medium);
+}
+
+@media screen and (max-width: 1199px) {
+  .author-section {
+    width: 181px;
+  }
+}
+
+@media screen and (max-width: 743px) {
+  .author-section {
+    position: static;
+
+    flex-direction: row;
+    gap: 62px;
+
+    width: 100%;
+
+    margin-bottom: 16px;
+    padding: 9px 16px;
+  }
+
+  .author-content {
+    font-size: 12px;
+    line-height: 18px;
+  }
+}
