+++ src/components/common/modal/detail-cards/CommentSection.module.css
@@ -1,0 +1,90 @@
+.comment-input-section {
+  position: relative;
+
+  width: 450px;
+}
+
+.comment-label {
+  width: 100%;
+
+  font-size: 16px;
+  font-weight: 500;
+  line-height: 26px;
+  color: var(--black-medium);
+}
+
+.comment-input {
+  display: block;
+
+  width: 100%;
+
+  margin-top: 4px;
+  padding: 16px;
+
+  border: 1px solid var(--gray-medium);
+  border-radius: 6px;
+
+  font-size: 14px;
+  line-height: 24px;
+
+  resize: none;
+}
+
+.btn-add-comment {
+  position: absolute;
+  right: 12px;
+  bottom: 12px;
+
+  width: 83px;
+  height: 32px;
+
+  border: 1px solid var(--gray-medium);
+  border-radius: 4px;
+  background-color: var(--white);
+
+  font-size: 12px;
+  font-weight: 500;
+  line-height: 18px;
+  color: var(--violet);
+
+  cursor: pointer;
+}
+
+.btn-add-comment:disabled {
+  background-color: var(--gray-bright);
+
+  color: var(--gray);
+
+  cursor: not-allowed;
+}
+
+.end-point {
+  height: 1px;
+}
+
+@media screen and (max-width: 1199px) {
+  .comment-input-section {
+    width: 100%;
+  }
+
+  .btn-add-comment {
+    right: 11px;
+
+    width: 77px;
+  }
+}
+
+@media screen and (max-width: 743px) {
+  .comment-input {
+    font-size: 12px;
+    line-height: 18px;
+  }
+
+  .btn-add-comment {
+    right: 12px;
+    bottom: 20px;
+
+    width: 84px;
+    height: 28px;
+  }
+}
