+++ src/components/common/modal/detail-cards/DetailCardModal.module.css
@@ -47,37 +47,6 @@
   height: 32px;
 }
 
-.author-section {
-  position: absolute;
-  top: 80px;
-  right: 32px;
-
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-
-  width: 200px;
-
-  padding: 14px 16px;
-
-  border: 1px solid var(--gray-medium);
-  border-radius: 8px;
-}
-
-.author-title {
-  margin-bottom: 6px;
-
-  font-size: 12px;
-  font-weight: 600;
-  line-height: 20px;
-}
-
-.author-content {
-  font-size: 14px;
-  line-height: 24px;
-  color: var(--black-medium);
-}
-
 .content-section {
   overflow: scroll;
 
@@ -87,34 +56,6 @@
 
 .content-section::-webkit-scrollbar {
   display: none;
-}
-
-.chip-section {
-  display: flex;
-  align-items: center;
-
-  margin-bottom: 20px;
-}
-
-.status {
-  flex-shrink: 0;
-
-  width: 63px;
-}
-
-.bar {
-  width: 1px;
-  height: 20px;
-
-  margin: 0 20px;
-
-  background-color: var(--gray-medium);
-}
-
-.tags {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 6px;
 }
 
 .description {
@@ -128,70 +69,6 @@
 .card-image {
   width: 445px;
   height: 260px;
-}
-
-.comment-input-section {
-  position: relative;
-
-  width: 450px;
-}
-
-.comment-label {
-  width: 100%;
-
-  font-size: 16px;
-  font-weight: 500;
-  line-height: 26px;
-  color: var(--black-medium);
-}
-
-.comment-input {
-  display: block;
-
-  width: 100%;
-
-  margin-top: 4px;
-  padding: 16px;
-
-  border: 1px solid var(--gray-medium);
-  border-radius: 6px;
-
-  font-size: 14px;
-  line-height: 24px;
-
-  resize: none;
-}
-
-.btn-add-comment {
-  position: absolute;
-  right: 12px;
-  bottom: 12px;
-
-  width: 83px;
-  height: 32px;
-
-  border: 1px solid var(--gray-medium);
-  border-radius: 4px;
-  background-color: var(--white);
-
-  font-size: 12px;
-  font-weight: 500;
-  line-height: 18px;
-  color: var(--violet);
-
-  cursor: pointer;
-}
-
-.btn-add-comment:disabled {
-  background-color: var(--gray-bright);
-
-  color: var(--gray);
-
-  cursor: not-allowed;
-}
-
-.end-point {
-  height: 1px;
 }
 
 @media screen and (max-width: 1199px) {
@@ -210,17 +87,9 @@
     right: 32px;
   }
 
-  .author-section {
-    width: 181px;
-  }
-
   .content-section {
     width: 420px;
     height: calc(100% - 80px);
-  }
-
-  .chip-section {
-    margin-bottom: 16px;
   }
 
   .description {
@@ -233,16 +102,6 @@
     height: 246px;
 
     margin-bottom: 24px;
-  }
-
-  .comment-input-section {
-    width: 100%;
-  }
-
-  .btn-add-comment {
-    right: 11px;
-
-    width: 77px;
   }
 }
 
@@ -276,34 +135,9 @@
     height: 24px;
   }
 
-  .author-section {
-    position: static;
-
-    flex-direction: row;
-    gap: 62px;
-
-    width: 100%;
-
-    margin-bottom: 16px;
-    padding: 9px 16px;
-  }
-
-  .author-content {
-    font-size: 12px;
-    line-height: 18px;
-  }
-
   .content-section {
     width: 100%;
     height: calc(100% - 190px);
-  }
-
-  .bar {
-    margin: 0 12px;
-  }
-
-  .tags {
-    gap: 8px;
   }
 
   .description {
@@ -312,18 +146,5 @@
 
   .card-image {
     height: 168px;
-  }
-
-  .comment-input {
-    font-size: 12px;
-    line-height: 18px;
-  }
-
-  .btn-add-comment {
-    right: 12px;
-    bottom: 20px;
-
-    width: 84px;
-    height: 28px;
   }
 }
